fix station id to sid; redis timeout type long; exception return back code 5

+++ provider/src/main/java/com/ycb/wxxcx/provider/cache/RedisService.java
@@ -20,7 +20,7 @@ public class RedisService {
     @Autowired
     private RedisTemplate<String, String> template;
 
-    public void setKeyValueTimeout(String key, String value, Integer timeout) {
+    public void setKeyValueTimeout(String key, String value, Long timeout) {
         template.setValueSerializer(new GenericToStringSerializer<>(String.class));
         template.opsForValue().set(key, value, timeout, TimeUnit.SECONDS);
     }
